[#9] Fix: vm_entry 구조체 타입 변경 및 hash_delete 변경 (5 Fail)

+++ include/userprog/process.h
@@ -21,8 +21,8 @@ struct thread *get_child_process (int pid);
 struct vm_entry{
     struct file *f;
     off_t offset;
-    size_t read_bytes;
-    size_t zero_bytes;
+    uint32_t read_bytes;
+    uint32_t zero_bytes;
 };
 
 #endif /* userprog/process.h */
