+++ include/vm/file.h
@@ -8,9 +8,9 @@ enum vm_type;
 
 struct file_page {
 	struct file *file;
-	size_t offset;
-	size_t read_bytes;
-	size_t zero_bytes;
+	off_t offset;
+	uint32_t read_bytes;
+	uint32_t zero_bytes;
 };
 
 void vm_file_init (void);
